class-02/airflow-spark/requirements.txt
apache-airflow-providers-docker==4.4.3